# Improve performance by utilizing cache locality