@lm-konda 2019-01-28
performance change


before
root@DE10_NANO:~# ./lm_arm.elf input.npy output.npy 
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
ApplyThresholds,129652,59918,20346,8050,  sum:217.966ms
AveragePool,86,  sum:0.086ms
BatchNorm,4495,116,  sum:4.611ms
Convolution,160054,2119,  sum:162.173ms
ExtractImagePatches,3917,906,360,  sum:5.183ms
QTZ_linear_mid_tread_half,6883,  sum:6.883ms
QuantizedConv2D,6941,  sum:6.941ms
QuantizedConv2D_ApplyScalingFactor,204,  sum:0.204ms
ReLu,165,  sum:0.165ms
SoftMax,134,  sum:0.134ms
TotalInitTime,39112,  sum:39.112ms
TotalRunTime,728040,  sum:728.04ms
kn2row,158792,  sum:158.792ms
kn2row-1x1,2110,  sum:2.11ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,107323,2100,  sum:109.423ms
matrix_shift_add_f1,3612,  sum:3.612ms
matrix_shift_add_f2,47810,  sum:47.81ms
matrix_transpose (row_major),23,  sum:0.023ms
```
after
```
root@DE10_NANO:~# ./lm_arm.elf input.npy output.npy 
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
ApplyThresholds,129243,60693,18603,7948,  sum:216.487ms
AveragePool,86,  sum:0.086ms
BatchNorm,4985,119,  sum:5.104ms
Convolution,159949,2122,  sum:162.071ms
ExtractImagePatches,3917,908,358,  sum:5.183ms
QTZ_linear_mid_tread_half,6924,  sum:6.924ms
QuantizedConv2D,5750,  sum:5.75ms
QuantizedConv2D_ApplyScalingFactor,188,  sum:0.188ms
ReLu,163,  sum:0.163ms
SoftMax,130,  sum:0.13ms
TotalInitTime,38909,  sum:38.909ms
TotalRunTime,651070,  sum:651.07ms
kn2row,158685,  sum:158.685ms
kn2row-1x1,2113,  sum:2.113ms
kn2row-buf,4,  sum:0.004ms
matrix_multiplication,107506,2104,  sum:109.61ms
matrix_shift_add_f1,3564,  sum:3.564ms
matrix_shift_add_f2,47569,  sum:47.569ms
matrix_transpose (row_major),22,  sum:0.022ms
```